Refactor cleanup scripts to use env helper
