+++ scripts/fixSessions.js
@@ -1,35 +1,14 @@
 const fs = require('fs');
 const path = require('path');
-const dotenv = require('dotenv');
+const { load, resolveDir } = require('./env');
 
-// Charger les variables d'environnement depuis le .env racine
-dotenv.config({ path: path.resolve(__dirname, '../.env') });
-// Puis surcharge avec celles du backend si disponibles
-dotenv.config({ path: path.resolve(__dirname, '../backend/.env'), override: true });
+// Charge les variables d'environnement (.env racine puis backend)
+load();
 
-function isAbsolute(p) {
-  return path.isAbsolute(p) || /^[A-Za-z]:[\\/]/.test(p);
-}
-
-function normalize(p) {
-  return p.split(/[\\/]+/).join(path.sep);
-}
-
-function getDir(envVar, fallback) {
-  const value = process.env[envVar];
-  if (value) {
-    const norm = normalize(value);
-    return isAbsolute(value)
-      ? path.normalize(norm)
-      : path.resolve(__dirname, '../backend', norm);
-  }
-  return path.resolve(__dirname, '../backend', fallback);
-}
-
-const DATA_DIR = getDir('DATA_DIR', 'src/data');
+const DATA_DIR = resolveDir('DATA_DIR', 'src/data');
 const FILE = path.join(DATA_DIR, 'analytics.json');
 
-function load() {
+function loadFile() {
   try {
     return JSON.parse(fs.readFileSync(FILE, 'utf8'));
   } catch {
@@ -37,12 +16,12 @@ function load() {
   }
 }
 
-function save(data) {
+function saveFile(data) {
   fs.writeFileSync(FILE, JSON.stringify(data, null, 2));
 }
 
 function addMissingLogouts() {
-  const data = load();
+  const data = loadFile();
   if (!Array.isArray(data.sessions)) return;
 
   let changed = false;
@@ -64,7 +43,7 @@ function addMissingLogouts() {
   });
 
   if (changed) {
-    save(data);
+    saveFile(data);
     console.log('Analytics updated.');
   } else {
     console.log('No missing logouts found.');
